refactor(model): relocate query_utils to base/name_field_utils
transform_name_fields/uses_internal_name had no actual uploaded-track
dependency; move them to a neutral location so BaseModel/BaseManager
can eventually live in a shared package independent of
uploaded_track_mixin.

Pure rename, no behavior change.
