+++ api/model/base/name_field_utils.py
@@ -5,13 +5,14 @@ from django.db import models
 
 from api.model.field.AppCharField import AppCharField
 
-from .Fields import Fields
+NAME_PUBLIC = "name"
+NAME_INTERNAL = f"_{NAME_PUBLIC}"
 
 
 def uses_internal_name(model: type[models.Model]) -> bool:
     try:
-        field = model._meta.get_field(Fields.NAME_INTERNAL)
-        return isinstance(field, AppCharField) and field.db_column == Fields.NAME_PUBLIC
+        field = model._meta.get_field(NAME_INTERNAL)
+        return isinstance(field, AppCharField) and field.db_column == NAME_PUBLIC
     except FieldDoesNotExist:
         return False
 
@@ -27,16 +28,16 @@ def transform_name_fields(model: type[models.Model], **kwargs: Any) -> dict[str,
 
     for key, value in kwargs.items():
         # Handle direct name field references
-        if key == Fields.NAME_PUBLIC and uses_internal_name(model):
-            transformed[Fields.NAME_INTERNAL] = value
+        if key == NAME_PUBLIC and uses_internal_name(model):
+            transformed[NAME_INTERNAL] = value
         # Handle relationship traversals and lookups containing __name
-        elif "__" + Fields.NAME_PUBLIC in key:
+        elif "__" + NAME_PUBLIC in key:
             # Split on __name to preserve any lookups that come after
-            parts = key.split("__" + Fields.NAME_PUBLIC)
+            parts = key.split("__" + NAME_PUBLIC)
             if len(parts) == 2:
                 # parts[0] is the relationship path
                 # parts[1] is either empty or contains lookups (like __icontains)
-                transformed[parts[0] + "__" + Fields.NAME_INTERNAL + parts[1]] = value
+                transformed[parts[0] + "__" + NAME_INTERNAL + parts[1]] = value
             else:
                 transformed[key] = value
         else:
